Delete unused imports and make smaller updates
