+++ jhotdraw-core/src/main/java/org/jhotdraw/draw/SPILocator.java
@@ -1,4 +1,4 @@
-package org.jhotdraw.draw.actionLocator;
+package org.jhotdraw.draw;
 
 import java.util.*;
