+++ jhotdraw-core/src/main/java/org/jhotdraw/draw/action/arrange/ArrangeService.java
@@ -1,4 +1,4 @@
-package org.jhotdraw.draw.action.arrangeActions;
+package org.jhotdraw.draw.action.arrange;
 
 import org.jhotdraw.draw.DrawingEditor;
 import org.jhotdraw.draw.action.AbstractSelectedAction;
