+++ jhotdraw-core/src/main/java/org/jhotdraw/draw/action/arrange/BringToFrontAction.java
@@ -1,11 +1,11 @@
 /*
- * @(#)SendToBackAction.java
+ * @(#)BringToFrontAction.java
  *
  * Copyright (c) 2003-2008 The authors and contributors of JHotDraw.
  * You may not use, copy or modify this file, except in compliance with the
  * accompanying license terms.
  */
-package org.jhotdraw.draw.action.arrangeActions;
+package org.jhotdraw.draw.action.arrange;
 
 import dk.sdu.mmmi.featuretracer.lib.FeatureEntryPoint;
 import org.jhotdraw.draw.action.AbstractSelectedAction;
@@ -16,39 +16,38 @@ import org.jhotdraw.draw.*;
 import org.jhotdraw.util.ResourceBundleUtil;
 
 /**
- * SendToBackAction.
+ * ToFrontAction.
  *
  * @author Werner Randelshofer
  * @version $Id$
  */
-public class SendToBackAction extends AbstractSelectedAction implements ArrangeService {
+public class BringToFrontAction extends AbstractSelectedAction implements ArrangeService{
+
     private static final long serialVersionUID = 1L;
-    public static final String ID = "edit.sendToBack";
+    public static final String ID = "edit.bringToFront";
 
-    public SendToBackAction(){}
+    public BringToFrontAction() {}
 
-    @Override
-    public SendToBackAction createWithEditor(DrawingEditor editor) {
-        SendToBackAction action = new SendToBackAction(editor);
-        return action;
+    public BringToFrontAction createWithEditor(DrawingEditor editor) {
+        return new BringToFrontAction(editor);
     }
 
     /**
      * Creates a new instance.
      */
-    @FeatureEntryPoint(value = "sendToBackAction")
-    public SendToBackAction(DrawingEditor editor) {
+    @FeatureEntryPoint(value = "bringToFrontAction")
+    public BringToFrontAction(DrawingEditor editor) {
         super(editor);
         ResourceBundleUtil.getRResourceBundleUtilLabels().configureAction(this, ID);
         updateEnabledState();
     }
 
-    @FeatureEntryPoint(value = "sendToBack_actionPreformed")
+    @FeatureEntryPoint(value = "bringToFront_actionPreformed")
     @Override
     public void actionPerformed(java.awt.event.ActionEvent e) {
         final DrawingView view = getView();
         final LinkedList<Figure> figures = new LinkedList<>(view.getSelectedFigures());
-        sendToBack(view, figures);
+        bringToFront(view, figures);
         fireUndoableEditHappened(new AbstractUndoableEdit() {
             private static final long serialVersionUID = 1L;
 
@@ -60,21 +59,21 @@ public class SendToBackAction extends AbstractSelectedAction implements ArrangeS
             @Override
             public void redo() throws CannotRedoException {
                 super.redo();
-                SendToBackAction.sendToBack(view, figures);
+                BringToFrontAction.bringToFront(view, figures);
             }
 
             @Override
             public void undo() throws CannotUndoException {
                 super.undo();
-                BringToFrontAction.bringToFront(view, figures);
+                SendToBackAction.sendToBack(view, figures);
             }
         });
     }
 
-    public static void sendToBack(DrawingView view, Collection<Figure> figures) {
+    public static void bringToFront(DrawingView view, Collection<Figure> figures) {
         Drawing drawing = view.getDrawing();
         for (Figure figure : drawing.sort(figures)) {
-            drawing.sendToBack(figure);
+            drawing.bringToFront(figure);
         }
     }
