+++ jhotdraw-core/src/main/java/org/jhotdraw/draw/action/arrange/SendToBackAction.java
@@ -1,11 +1,11 @@
 /*
- * @(#)BringToFrontAction.java
+ * @(#)SendToBackAction.java
  *
  * Copyright (c) 2003-2008 The authors and contributors of JHotDraw.
  * You may not use, copy or modify this file, except in compliance with the
  * accompanying license terms.
  */
-package org.jhotdraw.draw.action.arrangeActions;
+package org.jhotdraw.draw.action.arrange;
 
 import dk.sdu.mmmi.featuretracer.lib.FeatureEntryPoint;
 import org.jhotdraw.draw.action.AbstractSelectedAction;
@@ -16,39 +16,38 @@ import org.jhotdraw.draw.*;
 import org.jhotdraw.util.ResourceBundleUtil;
 
 /**
- * ToFrontAction.
+ * SendToBackAction.
  *
  * @author Werner Randelshofer
  * @version $Id$
  */
-public class BringToFrontAction extends AbstractSelectedAction implements ArrangeService{
-
+public class SendToBackAction extends AbstractSelectedAction implements ArrangeService {
     private static final long serialVersionUID = 1L;
-    public static final String ID = "edit.bringToFront";
+    public static final String ID = "edit.sendToBack";
 
-    public BringToFrontAction() {}
+    public SendToBackAction(){}
 
-    public BringToFrontAction createWithEditor(DrawingEditor editor) {
-        BringToFrontAction action = new BringToFrontAction(editor);
-        return action;
+    @Override
+    public SendToBackAction createWithEditor(DrawingEditor editor) {
+        return new SendToBackAction(editor);
     }
 
     /**
      * Creates a new instance.
      */
-    @FeatureEntryPoint(value = "bringToFrontAction")
-    public BringToFrontAction(DrawingEditor editor) {
+    @FeatureEntryPoint(value = "sendToBackAction")
+    public SendToBackAction(DrawingEditor editor) {
         super(editor);
         ResourceBundleUtil.getRResourceBundleUtilLabels().configureAction(this, ID);
         updateEnabledState();
     }
 
-    @FeatureEntryPoint(value = "bringToFront_actionPreformed")
+    @FeatureEntryPoint(value = "sendToBack_actionPreformed")
     @Override
     public void actionPerformed(java.awt.event.ActionEvent e) {
         final DrawingView view = getView();
         final LinkedList<Figure> figures = new LinkedList<>(view.getSelectedFigures());
-        bringToFront(view, figures);
+        sendToBack(view, figures);
         fireUndoableEditHappened(new AbstractUndoableEdit() {
             private static final long serialVersionUID = 1L;
 
@@ -60,21 +59,21 @@ public class BringToFrontAction extends AbstractSelectedAction implements Arrang
             @Override
             public void redo() throws CannotRedoException {
                 super.redo();
-                BringToFrontAction.bringToFront(view, figures);
+                SendToBackAction.sendToBack(view, figures);
             }
 
             @Override
             public void undo() throws CannotUndoException {
                 super.undo();
-                SendToBackAction.sendToBack(view, figures);
+                BringToFrontAction.bringToFront(view, figures);
             }
         });
     }
 
-    public static void bringToFront(DrawingView view, Collection<Figure> figures) {
+    public static void sendToBack(DrawingView view, Collection<Figure> figures) {
         Drawing drawing = view.getDrawing();
         for (Figure figure : drawing.sort(figures)) {
-            drawing.bringToFront(figure);
+            drawing.sendToBack(figure);
         }
     }
