+++ jhotdraw-core/src/main/resources/META-INF/services/org.jhotdraw.draw.action.arrange.ArrangeService
@@ -1,0 +1,2 @@
+org.jhotdraw.draw.action.arrange.SendToBackAction
+org.jhotdraw.draw.action.arrange.BringToFrontAction
