--- jhotdraw-core/src/main/resources/META-INF/services/org.jhotdraw.draw.action.arrangeActions.ArrangeService
@@ -1,2 +1,0 @@
-org.jhotdraw.draw.action.arrangeActions.SendToBackAction
-org.jhotdraw.draw.action.arrangeActions.BringToFrontAction
